First Scraping Rendition
This created the main for running scraping as well as the ScrapeWebsite module. Currently able to scrape from worldometer by reading a list of country names from the text file "scrapecountries.txt". Stores output as Excel and JSON in a folder called "Covid_Data" with the files named as the date of the scrape. No special formatting or ordering has been performed on the output.